test(datetime): migrate multiple
